requirements.txt
accelerate==1.4.0
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diffusers==0.32.2
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
narwhals==1.27.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
streamlit==1.42.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
zipp==3.21.0
